Move selector specific tests to test_selector
